Use dust limit, set sane default tx fee, use minrelayfee as marker
1. `GetDustLimit(const CScript&)` is used instead of `MP_DUST_LIMIT`. See 426592e287f79b37fbdddb5cad14ce9bf2ca1317 for additional details.

2. `nMinRelayTxFee` was set to 1001 which results in output amounts of:
 - 552 (pay-to-pubkey-hash)
 - 685 (multisig, two compressed public keys)
 - 781 (multisig, one uncompressed, one compressed public key)
 - 787 (multisig, three compressed public keys)
 - 883 (multisig, one uncompressed, two compressed public keys)

Set this value to 10000 to mirror the old "6000 / packet rule" with slightly tighter (and more accurate) amounts.

The default value was set to 1000 back in November 2013, so this amount should be safe. +1 Satoshi is used as marker. Increment further, if a new marker is needed.

The `minrelaytxfee` can be set via bitcoin.conf or startup parameter. Note: `minrelaytxfee=0.00001001` in decimal form must used! Verify via `getinfo` and look out for `relayfee`.

3. `DEFAULT_TRANSACTION_FEE` was set to 10000 from 0.

This simply mirrors the actual default value and has no effect, unless the mintxfee is overwritten. In this case all it does is to act as safeguard for the case that only `mintxfee` was set and `paytxfee` was forgotten.