Make sure LoadFreeze() uses actual position in blockchain
Note: during some testing I picked up that while loading the freeze state at startup if there are multiple freeze/unfreeze transactions in the same block they may not be loaded in the correct order.  Unlikely scanerio but we need to cover it just the same.  This change just adds the tx position within the block to the sort order.